feat: Add the optional deserializer
The optionalDeserializer allows to handle the case in which the data is null returning the None.

Ref #37
